requirements.txt
jinja2
pyyaml
docker
matplotlib